pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.2.0",
    "django>=5.1.1",
    "django-allauth>=65.0.2",
    "django-ninja>=1.3.0",
    "djecorator>=1.2.0",
    "granian>=1.6.0",
    "psycopg2-binary>=2.9.9",
    "python-decouple>=3.8",
    "django-cors-headers>=4.4.0",
    "django-anymail[amazon-ses]>=12.0",
    "loguru>=0.7.3",
    "django-browser-reload>=1.17.0",
    "sentry-sdk>=2.20.0",
    "whitenoise>=6.8.2",
    "django-widget-tweaks>=1.5.0",
    "django-debug-toolbar>=5.0.1",
    "django-redis>=5.4.0",
    "django-vite>=3.0.2",
    "dbos>=1.12.0",
]

[dependency-groups]
dev = [
    "blessed>=1.20.0",
    "flower>=2.0.1",
    "setproctitle>=1.3.5",
    "textual>=3.0.0",
]
